[Signal Generator] Change the application's power supply
In order to achieve high precision in high speed data conversion
applications, the analog circuitry should not share a common power
supply rail and a common ground as the digital circuitry of the project.
As such, there are separate +5V rails for the digital circuitry (+5VD)
and for the analog circuitry (+5VA), as well as different grounds
(respectively GNDD and GNDA).
Furthermore, there are plans to use OpAmps in the DAC's output circuitry
which are to be powered by the +5VA/-5VA rails and the +12VA/-12VA
rails.

The current schematic's power connections have been altered to implement
the explained power supply changes above.