[Signal generator] Connect AGND and DGND in a star point
In a single PCB mixed-signal system, the analog and digital grounds
should be connected to each other at only one point, which is called the
'star' point.

Note: the star connection has been made in KiCad with a wire which
causes the ERC to produce an error. The latter should be ignored.