[Power Supply] Start design of temporary supply. Design analog supply
In order to program and troubleshoot the signal generator board, a
temporary linear power supply is to be designed. It is meant to be easy
to construct, having common THT parts.

Currently, the analog power supply, which provides +12V/-12V and +5V/-5V
rails (to AGND), has been designed. It is separated in its own hierarchical
sheet.