[Power Supply] Finalize design of temporary power supply
This temporary power supply outputs the following rails:
- +12V/-12V analog rails to AGND. These rails are connected to the
Analog_Supply_Rails connector (J1);
- +5V/-5V analog rails to AGND. These rails are connected to the
Analog_Supply_Rails connector (J1);
- +5V digital rail to DGND. This rails is outsourced to the
Digital_Supply_Rails connector (J2).
No PCB is planned to be implemented for this linear power supply. This
is why footprints will not be assigned to components.
Note: this is not the final power supply module for this project. The
current power supply is to be used for testing and troubleshooting
purposes.

All remnants of the previous power supply design have been removed.