[Documentation] Start work on BG project documentation
The project itself is to be presented as a prototype since it is not
complete (needs proper power supply, software implementation, etc.). All
of the missing functionality will be described in the section for future
improvements.

As of now, only the first page and the 'Contents' page have been
written.

Git LFS is used to track PDF files, such as this project documentation.